+++ Core/OfficeDevPnP.Core/Extensions/TenantExtensions.cs
@@ -22,6 +22,7 @@ using OfficeDevPnP.Core.Framework.Provisioning.Model;
 using OfficeDevPnP.Core.Framework.Provisioning.ObjectHandlers;
 using OfficeDevPnP.Core.Utilities;
 using Newtonsoft.Json.Linq;
+using OfficeDevPnP.Core.Framework.Provisioning.Model.Configuration;
 #endif
 
 namespace Microsoft.SharePoint.Client
@@ -36,10 +37,69 @@ namespace Microsoft.SharePoint.Client
 #if !ONPREMISES
         #region Provisioning
 
-        public static void ApplyProvisionHierarchy(this Tenant tenant, ProvisioningHierarchy hierarchy, string sequenceId, ProvisioningTemplateApplyingInformation applyingInformation = null)
+        /// <summary>
+        /// Applies a template to a tenant
+        /// </summary>
+        /// <param name="tenant"></param>
+        /// <param name="tenantTemplate"></param>
+        /// <param name="sequenceId"></param>
+        /// <param name="configuration"></param>
+        public static void ApplyTenantTemplate(this Tenant tenant, ProvisioningHierarchy tenantTemplate, string sequenceId, ApplyConfiguration configuration = null)
         {
             SiteToTemplateConversion engine = new SiteToTemplateConversion();
-            engine.ApplyProvisioningHierarchy(tenant, hierarchy, sequenceId, applyingInformation);
+            engine.ApplyTenantTemplate(tenant, tenantTemplate, sequenceId, configuration);
+        }
+
+        /// <summary>
+        /// Extracts a template from a tenant
+        /// </summary>
+        /// <param name="tenant"></param>
+        /// <param name="configuration"></param>
+        /// <returns></returns>
+        public static ProvisioningHierarchy GetTenantTemplate(this Tenant tenant, ExtractConfiguration configuration)
+        {
+            return new SiteToTemplateConversion().GetTenantTemplate(tenant, configuration);
+        }
+
+        /// <summary>
+        /// Returns the urls of sites connected to the hubsite specified
+        /// </summary>
+        /// <param name="tenant">A tenant object pointing to the context of a Tenant Administration site</param>
+        /// <param name="hubSiteUrl">The fully qualified url of the hubsite</param>
+        /// <returns></returns>
+        public static List<string> GetHubSiteChildUrls(this Tenant tenant, string hubSiteUrl)
+        {
+            var properties = tenant.GetHubSitePropertiesByUrl(hubSiteUrl);
+            tenant.Context.Load(properties);
+            tenant.Context.ExecuteQueryRetry();
+            return GetHubSiteChildUrls(tenant, properties.ID);
+        }
+
+        /// <summary>
+        /// Returns the urls of sites connected to the hubsite specified
+        /// </summary>
+        /// <param name="tenant">A tenant object pointing to the context of a Tenant Administration site</param>
+        /// <param name="hubsiteId">The id of the hubsite</param>
+        /// <returns></returns>
+        public static List<string> GetHubSiteChildUrls(this Tenant tenant, Guid hubsiteId)
+        {
+            List<string> urls = new List<string>();
+            using (var tenantContext = tenant.Context.Clone((tenant.Context as ClientContext).Web.GetTenantAdministrationUrl()))
+            {
+                var siteList = tenantContext.Web.Lists.GetByTitle("DO_NOT_DELETE_SPLIST_TENANTADMIN_AGGREGATED_SITECOLLECTIONS");
+                var query = new CamlQuery()
+                {
+                    ViewXml = $"<View><Query><Where><And><Eq><FieldRef Name='HubSiteId' /><Value Type='Guid'>{hubsiteId}</Value></Eq><And><Neq><FieldRef Name='SiteId' /><Value Type='Guid'>{hubsiteId}</Value></Neq><IsNull><FieldRef Name='TimeDeleted'/></IsNull></And></And></Where></Query><ViewFields><FieldRef Name='SiteUrl'/></ViewFields></View>"
+                };
+                var items = siteList.GetItems(query);
+                tenantContext.Load(items);
+                tenantContext.ExecuteQueryRetry();
+                foreach (var item in items)
+                {
+                    urls.Add(item["SiteUrl"].ToString());
+                }
+            }
+            return urls;
         }
         #endregion
 
@@ -116,7 +176,7 @@ namespace Microsoft.SharePoint.Client
         }
 
         /// <summary>
-        /// Launches a site collection creation and waits for the creation to finish 
+        /// Launches a site collection creation and waits for the creation to finish
         /// </summary>
         /// <param name="tenant">A tenant object pointing to the context of a Tenant Administration site</param>
         /// <param name="siteFullUrl">The SPO URL</param>
@@ -157,7 +217,7 @@ namespace Microsoft.SharePoint.Client
 
         #region Site status checks
         /// <summary>
-        /// Returns if a site collection is in a particular status. If the URL contains a sub site then returns true is the sub site exists, false if not. 
+        /// Returns if a site collection is in a particular status. If the URL contains a sub site then returns true is the sub site exists, false if not.
         /// Status is irrelevant for sub sites
         /// </summary>
         /// <param name="tenant">A tenant object pointing to the context of a Tenant Administration site</param>
@@ -244,12 +304,12 @@ namespace Microsoft.SharePoint.Client
         }
 
         /// <summary>
-        /// Checks if a site collection exists, relies on tenant admin API. Sites that are recycled also return as existing sites
+        /// Checks if a site collection exists, relies on tenant admin API. Sites that are recycled also return as existing sites, but with a different flag
         /// </summary>
         /// <param name="tenant">A tenant object pointing to the context of a Tenant Administration site</param>
         /// <param name="siteFullUrl">URL to the site collection</param>
-        /// <returns>True if existing, false if not</returns>
-        public static bool SiteExists(this Tenant tenant, string siteFullUrl)
+        /// <returns>An enumerated type that can be: No, Yes, Recycled</returns>
+        public static SiteExistence SiteExistsAnywhere(this Tenant tenant, string siteFullUrl)
         {
             try
             {
@@ -259,7 +319,7 @@ namespace Microsoft.SharePoint.Client
                 tenant.Context.ExecuteQueryRetry();
 
                 // Will cause an exception if site URL is not there. Not optimal, but the way it works.
-                return true;
+                return SiteExistence.Yes;
             }
             catch (Exception ex)
             {
@@ -273,21 +333,28 @@ namespace Microsoft.SharePoint.Client
                             var deletedProperties = tenant.GetDeletedSitePropertiesByUrl(siteFullUrl);
                             tenant.Context.Load(deletedProperties);
                             tenant.Context.ExecuteQueryRetry();
-                            return deletedProperties.Status.Equals("Recycled", StringComparison.OrdinalIgnoreCase);
+                            if (deletedProperties.Status.Equals("Recycled", StringComparison.OrdinalIgnoreCase))
+                            {
+                                return SiteExistence.Recycled;
+                            }
+                            else
+                            {
+                                return SiteExistence.No;
+                            }
                         }
                         catch
                         {
-                            return false;
+                            return SiteExistence.No;
                         }
                     }
                     else
                     {
-                        return false;
+                        return SiteExistence.No;
                     }
                 }
                 else
                 {
-                    return true;
+                    return SiteExistence.Yes;
                 }
             }
         }
@@ -346,7 +413,7 @@ namespace Microsoft.SharePoint.Client
             {
                 if (!useRecycleBin && IsCannotRemoveSiteException(ex))
                 {
-                    //eat exception as the site might be in the recycle bin and we allowed deletion from recycle bin 
+                    //eat exception as the site might be in the recycle bin and we allowed deletion from recycle bin
                 }
                 else
                 {
@@ -558,7 +625,7 @@ namespace Microsoft.SharePoint.Client
         /// <param name="tenant">A tenant object pointing to the context of a Tenant Administration site (i.e. https://[tenant]-admin.sharepoint.com)</param>
         /// <param name="siteFullUrl">The target site to change the lock state.</param>
         /// <param name="lockState">The target state the site should be changed to.</param>
-        /// <param name="wait">If true, processing will halt until the site collection lock state has been implemented</param>      
+        /// <param name="wait">If true, processing will halt until the site collection lock state has been implemented</param>
         /// <param name="timeoutFunction">An optional function that will be called while waiting for the site to be created. If set will override the wait variable. Return true to cancel the wait loop.</param>
         public static void SetSiteLockState(this Tenant tenant, string siteFullUrl, SiteLockState lockState, bool wait = false, Func<TenantOperationMessage, bool> timeoutFunction = null)
         {
@@ -628,7 +695,7 @@ namespace Microsoft.SharePoint.Client
 #if ONPREMISES
 #if !SP2013
         /// <summary>
-        /// Returns all site collections in the current Tenant based on a startIndex. IncludeDetail adds additional properties to the SPSite object. 
+        /// Returns all site collections in the current Tenant based on a startIndex. IncludeDetail adds additional properties to the SPSite object.
         /// </summary>
         /// <param name="tenant">Tenant object to operate against</param>
         /// <param name="startIndex">Start getting site collections from this index. Defaults to 0</param>
@@ -672,7 +739,7 @@ namespace Microsoft.SharePoint.Client
 #endif
 #else
         /// <summary>
-        /// Returns all site collections in the current Tenant based on a startIndex. IncludeDetail adds additional properties to the SPSite object. 
+        /// Returns all site collections in the current Tenant based on a startIndex. IncludeDetail adds additional properties to the SPSite object.
         /// </summary>
         /// <param name="tenant">Tenant object to operate against</param>
         /// <param name="startIndex">Not relevant anymore</param>
@@ -901,7 +968,7 @@ namespace Microsoft.SharePoint.Client
         #region Site Classification configuration
 
         /// <summary>
-        /// Enables Site Classifications for the target tenant 
+        /// Enables Site Classifications for the target tenant
         /// </summary>
         /// <param name="tenant">The target tenant</param>
         /// <param name="accessToken">The OAuth accessToken for Microsoft Graph with Azure AD</param>
@@ -912,7 +979,7 @@ namespace Microsoft.SharePoint.Client
         }
 
         /// <summary>
-        /// Enables Site Classifications for the target tenant 
+        /// Enables Site Classifications for the target tenant
         /// </summary>
         /// <param name="tenant">The target tenant</param>
         /// <param name="accessToken">The OAuth accessToken for Microsoft Graph with Azure AD</param>
@@ -925,7 +992,7 @@ namespace Microsoft.SharePoint.Client
         }
 
         /// <summary>
-        /// Enables Site Classifications for the target tenant 
+        /// Enables Site Classifications for the target tenant
         /// </summary>
         /// <param name="tenant">The target tenant</param>
         /// <param name="accessToken">The OAuth accessToken for Microsoft Graph with Azure AD</param>
@@ -1057,7 +1124,7 @@ namespace Microsoft.SharePoint.Client
                         Tenant tenant = new Tenant(adminContext);
                         tenant.EnsureProperty(t => t.RootSiteUrl);
 
-                        // If we've got access to the tenant admin context, 
+                        // If we've got access to the tenant admin context,
                         // it means that the currently connecte user is an admin
                         return (true);
                     }
@@ -1164,7 +1231,7 @@ namespace Microsoft.SharePoint.Client
         }
         #endregion
 
-#region Utilities
+        #region Utilities
 
 #if !ONPREMISES
         public static string GetTenantIdByUrl(string tenantUrl)
@@ -1204,7 +1271,26 @@ namespace Microsoft.SharePoint.Client
             return index != -1 ? originalString.Substring(prefix.Length, index - prefix.Length) : null;
         }
 
-#endregion
+        #endregion
 
+    }
+
+    /// <summary>
+    /// Defines the existence status of a Site Collection
+    /// </summary>
+    public enum SiteExistence
+    {
+        /// <summary>
+        /// The Site Collection does not exist
+        /// </summary>
+        No,
+        /// <summary>
+        /// The Site Collection exists
+        /// </summary>
+        Yes,
+        /// <summary>
+        /// The Site Collection is in the Recycle Bin
+        /// </summary>
+        Recycled,
     }
 }